+++ src/config-loader.ts
@@ -2,6 +2,7 @@ import { existsSync } from 'fs';
 import { basename, resolve } from 'path';
 import * as deepmerge from 'deepmerge';
 import { IScript, IScripts, Scripts } from './scripts';
+import * as path from 'path';
 import glob = require('glob');
 
 export interface IMenu {
@@ -19,6 +20,7 @@ interface IOptions {
   menu: {
     defaultScript: IScript;
     defaultChoice: string;
+    timeout: number;
     pageSize: number;
   };
 }
@@ -70,6 +72,7 @@ export class Config {
       menu: {
         defaultChoice: '',
         defaultScript: '',
+        timeout: 0,
         pageSize: 7,
       },
     },
@@ -133,7 +136,7 @@ export class Config {
 
   };
 
-  public static load(): Config {
+  public static load(directory: string): Config {
     const hash = new Set<string>();
     let config = new Config(Config.default);
     let files = Config.default.options.files;
@@ -144,7 +147,9 @@ export class Config {
 
       for (const file of files) {
         if (file && !hash.has(file)) {
-          if (existsSync(resolve(file))) config = config.merge(file);
+          const fullPath = path.join(directory, file);
+
+          if (existsSync(resolve(fullPath))) config = config.merge(fullPath);
 
           hash.add(file);